kagami/requirements.txt
aiohttp==3.10.5
aiosignal==1.3.1
aiosqlite==0.20.0
attrs==24.2.0
discord.py==2.4.0
frozenlist==1.4.1
idna==3.8
multidict==6.0.5
Pillow==10.1.0
python-dotenv==1.0.1
typing_extensions==4.12.2
wavelink==2.6.5
yarl==1.8.2